+++ lib/l10n/app_localizations_cs.dart
@@ -3102,6 +3102,9 @@ class AppLocalizationsCs extends AppLocalizations {
       'Nebyl nahrán žádný zvuk. Zkus to znovu.';
 
   @override
+  String get chatInputRecordingFailed => 'Nahrávání selhalo. Zkus to znovu.';
+
+  @override
   String get chatInputPleaseWait => 'Čekejte prosím...';
 
   @override