+++ lib/l10n/app_localizations_de.dart
@@ -3086,6 +3086,10 @@ class AppLocalizationsDe extends AppLocalizations {
       'Es wurde kein Audio aufgenommen. Versuch es noch einmal.';
 
   @override
+  String get chatInputRecordingFailed =>
+      'Aufnahme fehlgeschlagen. Bitte erneut versuchen.';
+
+  @override
   String get chatInputPleaseWait => 'Bitte warten...';
 
   @override