+++ lib/l10n/app_localizations_en.dart
@@ -3046,6 +3046,9 @@ class AppLocalizationsEn extends AppLocalizations {
   String get chatInputNoAudioRecorded => 'No audio was recorded. Try again.';
 
   @override
+  String get chatInputRecordingFailed => 'Recording failed. Please try again.';
+
+  @override
   String get chatInputPleaseWait => 'Please wait...';
 
   @override