+++ lib/l10n/app_localizations_it.dart
@@ -3102,6 +3102,9 @@ class AppLocalizationsIt extends AppLocalizations {
   String get chatInputNoAudioRecorded => 'No audio was recorded. Try again.';
 
   @override
+  String get chatInputRecordingFailed => 'Registrazione fallita. Riprova.';
+
+  @override
   String get chatInputPleaseWait => 'Ti prego, aspetta...';
 
   @override